backend/requirements.txt
Flask
openai
boto3
python-nmap
requests
beautifulsoup4
flask-cors